requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
Pillow==9.5.0
tk==0.1.0
Werkzeug==2.2.3
